Get architecture from PE files
The Linux kernel uses the PE format for the binary files of aarch64
architecture. Firecracker also expects the same format to load a kernel.
Therefore, unikernels running on top of Firecracker in aarch64 might use
the PE format too. As a result, we add support for retrieving the
architecture of a PE binary file.

Signed-off-by: Charalampos Mainas <cmainas@Nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
PR: #33