Recognize Linux kernel ARM64 boot executable Image
Some unikernels are using a specific binary format for aarch64 binaries.
An example of such unikernels is Unikraft. The used format is neither an
elf nor a PE, but can be recognized through its magic as a Linux kernel
ARM64 boot executable Image. For that purpose, we add support to
recognize this binary format.

Signed-off-by: Charalampos Mainas <cmainas@Nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
PR: #33